+++ src/ui/panes/lyrics.rs
@@ -60,7 +60,7 @@ impl Pane for LyricsPane {
             default_style
         };
 
-        let show_timestamp = context.config.theme.lyrics.show_timestamp;
+        let timestamp = context.config.theme.lyrics.timestamp;
 
         let mut current_area = middle_row as usize;
         let Some(current_line) = lrc.lines.get(current_line_idx) else {
@@ -72,7 +72,7 @@ impl Pane for LyricsPane {
             let Some(area) = areas.get(current_area) else {
                 break;
             };
-            let text = Text::from(if index == 0 && show_timestamp && !l.is_empty() {
+            let text = Text::from(if index == 0 && timestamp && !l.is_empty() {
                 Cow::Owned(format!("[{}] {}", current_line.time.to_string(), l))
             } else {
                 l
@@ -93,13 +93,13 @@ impl Pane for LyricsPane {
             for (index, l) in
                 textwrap::wrap(&line.content, area.width as usize).into_iter().enumerate().rev()
             {
-                let Some(area) = areas.get(before_area_cursor - 1) else {
-                    break;
-                };
                 if before_area_cursor == 0 {
                     break;
                 }
-                let text = Text::from(if index == 0 && show_timestamp && !l.is_empty() {
+                let Some(area) = areas.get(before_area_cursor - 1) else {
+                    break;
+                };
+                let text = Text::from(if index == 0 && timestamp && !l.is_empty() {
                     Cow::Owned(format!("[{}] {}", line.time.to_string(), l))
                 } else {
                     l
@@ -128,7 +128,7 @@ impl Pane for LyricsPane {
                 let Some(area) = areas.get(after_area_cursor + 1) else {
                     break;
                 };
-                let text = Text::from(if index == 0 && show_timestamp && !l.is_empty() {
+                let text = Text::from(if index == 0 && timestamp && !l.is_empty() {
                     Cow::Owned(format!("[{}] {}", line.time.to_string(), l))
                 } else {
                     l